# openapi/generated_openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigPayload": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          }
        },
        "required": [
          "config"
        ],
        "title": "ConfigPayload",
        "type": "object"
      },
...
      "GenerateCheckPayload": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "mode": {
            "default": "mvp",
            "enum": [
              "mvp",
              "strict"
            ],
            "title": "Mode",
            "type": "string"
          },
          "overwrite_policy": {
            "const": "managed",
            "default": "managed",
            "title": "Overwrite Policy",
            "type": "string"
          }
        },
        "title": "GenerateCheckPayload",
        "type": "object"
      },
      "GeneratePayload": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "export_dir": {
            "default": "exports",
            "title": "Export Dir",
            "type": "string"
          },
          "export_enabled": {
            "default": true,
            "title": "Export Enabled",
            "type": "boolean"
          },
          "mode": {
            "default": "mvp",
            "enum": [
              "mvp",
              "strict"
            ],
            "title": "Mode",
            "type": "string"
          },
          "overwrite_policy": {
            "const": "managed",
            "default": "managed",
            "title": "Overwrite Policy",
            "type": "string"
          }
        },
        "title": "GeneratePayload",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SavePayload": {
        "properties": {
          "apply_runtime": {
            "default": true,
            "title": "Apply Runtime",
            "type": "boolean"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          }
        },
        "required": [
          "config"
        ],
        "title": "SavePayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Connecting specialty grain producers with global buyers",
    "title": "GrainPlaza",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/setup/generate": {
      "post": {
        "operationId": "generate_project_api_setup_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Project Api Setup Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Project",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/generate/check": {
      "post": {
        "operationId": "check_generated_sync_api_setup_generate_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCheckPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Check Generated Sync Api Setup Generate Check Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Generated Sync",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/render-yaml": {
      "post": {
        "operationId": "render_yaml_api_setup_render_yaml_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Render Yaml Api Setup Render Yaml Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render Yaml",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/save": {
      "post": {
        "operationId": "save_config_api_setup_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Config Api Setup Save Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Config",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/validate": {
      "post": {
        "operationId": "validate_config_api_setup_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Config Api Setup Validate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Config",
        "tags": [
          "setup"
        ]
      }
    },
...
  }
}